refactor: use discovery configuration
To safely support arbitrary devices, we should do our best to only connect to known devices.

This commit introduces a generic config, modeled after the one for buttplug.io.

We do our scanning discovery as usual, but only connect to devices that we have explicitly
decided to support in the config, and only on channels that we've pre-identified as good.

This also allows us more easily map identified devices to a speific implmentation driver.